refactor: Fix warnings & add logging for exceptions instead of printStackTrace()

+++ src/main/java/me/alpha432/oyvey/features/gui/items/buttons/Slider.java
@@ -68,8 +68,8 @@ public class Slider
             double result = (Double) this.setting.getMin() + (double) ((float) this.difference * percent);
             this.setting.setValue((double) Math.round(10.0 * result) / 10.0);
         } else if (this.setting.getValue() instanceof Float) {
-            float result = ((Float) this.setting.getMin()).floatValue() + (float) this.difference * percent;
-            this.setting.setValue(Float.valueOf((float) Math.round(10.0f * result) / 10.0f));
+            float result = this.setting.getMin().floatValue() + (float) this.difference * percent;
+            this.setting.setValue((float) Math.round(10.0f * result) / 10.0f);
         } else if (this.setting.getValue() instanceof Integer) {
             this.setting.setValue((Integer) this.setting.getMin() + (int) ((float) this.difference * percent));
         }
